implemented rectangle methods

+++ TestSharpShapes/UnitTestShape.cs
@@ -12,7 +12,7 @@ namespace TestSharpShapes
             throw new NotImplementedException();
         }
 
-        public override int Perimeter()
+        public override decimal Perimeter()
         {
             throw new NotImplementedException();
         }
